pyproject.toml
[tool.poetry]
name = "Cadet"
version = "0.1.0"
description = "A tool for build new spaCy Language objects."
authors = ["apjanco <apjanco@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
spacy = "^3.7.2"
spacy-lookups-data = "^1.0.5"
uvicorn = "^0.25.0"
python-multipart = "^0.0.6"
httpx = "^0.26.0"
python-slugify = "^8.0.1"
dkpro-cassis = "^0.8.0"
aiofiles = "^23.2.1"
pytz = "^2023.3.post1"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl"}
xx-ent-wiki-sm = {url = "https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.7.0/xx_ent_wiki_sm-3.7.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
